move subtraction by one out of prompt_pure_string_length
prompt_pure_string_length is used for calculating the length of the
preprompt, but the preprompt doesn't contain a newline, so we shouldn't
subtract by one in prompt_pure_string_length.

What should happen instead is that when we subtract one from the length
when we calculate the number of lines the preprompt occupies.