Update readme, pre-emptive FAQ re: clear-screen.
Because some external zsh modules take control of your zsh widgets in an
unobvious way, the user needs to know what is happening and how to fix
it. I think it's better to let it break our custom clearing and let the
user know why it boke.

Trying to do the same as zsh-syntax-hilighting might result in odd cases
where zsh-syntax-highlighting registers the builtin clear-screen widget
and pure wrapping the custom highlighting widget resulting in multiple
clear screens.