Attempt to detect user initiated git fetch
When a user initiated git fetch is detected, all current async jobs for
pure are aborted in an attempt to give way for the user initiated one.

Without this, there is always a chance that pure is running a git fetch
at the same time as the user, essentially preventing the user from
updating some objects.