Require zsh 5.2+
We know pure works with older versions of zsh, but there are many
problems there were not fixed until zsh 5.1.1. Zsh 5.2 provides the
best user experience by not requiring kill signals to indicate the
completion of async tasks.

For the brave, pure can work on older versions, but we do not provide
support for them.