Check and set `prompt_opts` when not using `promptinit` (#277)
The `prompt_opts` array only has an effect when the theme is loaded via
promptinit (`prompt pure`). Previously they were not set when pure was
loaded via `source /path/to/pure.zsh` as is the case with e.g. antigen,
antibody, etc.

This commit attempts to detect whenever pure is not autoloaded via
promptinit and manually sets the options in `prompt_opts`.

Fixes #276.