Fix line erasure when terminal output does not end in newline
It seems that adding the initial newline in `$PROMPT` allows ZLE to
erase the entile line where the newline began. This is problematic when
command output does not end with its own newline. A subsequent terminal
resize or prompt update that triggers a prompt redraw will then erase
the line.

To fix this, we separate the newline from the prompt by calling print
manually during precmd.

Previously, Pure tried to keep any potential prefixes the user might
have added to the PROMPT (before preprompt). This fix no longer allows
such prefixes as doing so would be hackish, and likely a recipe for
other unforseen behavior.

NOTE: Because the newline is printed by the Pure precmd hook, it's
possible that another precmd hook is run afterwards, and if that hook
outputs any text, the newline will appear in the wrong place. A possible
solution would be to delay adding the hook or at a later point sort the
hooks so that Pure is last.

Fixes #390, #376.