Try to detech SSH connection when SSH_CONNECTION is unset (#393)
This change allows us to detect SSH connections even when the
`$SSH_CONNECTION` environment variable is gone. This can happen when a
user changes to another user (`su`, `sudo`) or the environment is
otherwise reset.

By checking the output of `who am i` for a parenthesis at the end, we
assume that said parenthesis must contain either an IPv4 or IPv6
address, indicating that this is a remote session.

The reliablity of this assumption has not been analyzed to a great
degree, there could be false positives, or systems that do not format
the output in this way.

Fixes #382.