Set title via atomic print statement (#399)
Previously, `setopt xtrace` would produce terminal bell rings (by \a)
whenever the title was set because of debug output between print
statements. As a result, the title was not set and the terminal bell
rang.

This change also utilizes the improved ssh connection detection to
provide hostname in title when changing user on remote machines.