Add filename to debug prompt
This commit also adds the filename (without path) to the debug prompt,
when applicable. It also makes sure to avoid displaying duplicate
information, say, when a file is being parsed, both %N and %x contain
the same value (file path).

There is a slight cost to doing the extra processing when testing in
iTerm (unscientific benchmark):

	zsh -x  0.30s user 0.14s system 57% cpu 0.754 total

vs

	zsh -x  0.49s user 0.16s system 65% cpu 0.988 total

I think this is within an acceptable range.