Improve the debug prompt (PS4) (#396)
* Improve the debug prompt (PS4)

Add colors, to highlight the current function and de-empasize line numbers.

Add an extra line number (`%I`) which represents the line number in the file
where the code resides (by default, only the function line number is shown,
making it hard(er) to find the relevant code).

Repeat the + character to indicate the current execution depth, e.g. when a 
function calls another, the depth is increased. This allows you to quickly see 
which function invoked which.

* Add filename to debug prompt

Add filename (without path) to the debug prompt, when applicable and makes sure
to avoid displaying duplicate information. Say, when a file is being parsed,
both %N and %x contain the same value (file path). In this case, we only show
one value.

There is a slight cost to doing the extra processing when testing in
iTerm (unscientific benchmark):

	zsh -x  0.30s user 0.14s system 57% cpu 0.754 total

vs

	zsh -x  0.49s user 0.16s system 65% cpu 0.988 total

I think this is within an acceptable range.