Update to zsh-async 1.7.1 and recover from unexpected worker death (#454)
With the updated version of zsh-async we can now get a signal that the zsh-async
worker has dies (this can happen in some rare cases). As a result, we can now
revive the worker by re-initallizing the async worker on next prompt.

Zsh-async also handles the death of the worker more gracefully now, even if we
don't react to the signal.